refactor(#3479): extract state extractors from recovery_engine.rs
Behavior-preserving (verbatim) extraction of the inflight-state derivation
helper cluster out of recovery_engine.rs into a new sibling leaf module
recovery_engine/state_extractors.rs.

Moved (14 helpers, ~183 prod LoC): the interrupted/stale handoff message
builder, the ready-for-input probes (tui-pane + legacy-tmux fallback), the
worktree path/branch/dispatch-id accessors, the tmux-session-name derivation,
the worktree-info reconstruction (git rev-parse + canonicalize), and the
spawn-cwd / session-worktree wiring those feed. Function bodies are
byte-identical; only `fn` visibility prefixes changed.

Wiring: child pulls parent items via `use super::*`. Root-called helpers are
re-imported byte-identical; `save_missing_session_handoff` (called from
recovery_paths::restart) is widened to `pub(in crate::services::discord)` and
re-exported so its `recovery_engine::save_missing_session_handoff` path stays
valid. Cluster-internal helpers stay private to the submodule.

Ratchet: recovery_engine.rs giant baseline 3607 -> 3424. New file is under the
1000-prod-LoC giant threshold (no registry entry needed). Regenerated
ARCHITECTURE.md / inventory docs and synced the change-surfaces freeze entry.
